Watchlist changes & ui changes
- watchlist impl
- ui changes for watchlist, login button [ image needs to be finalyzed ]
- guest login button

+++ src/main/webapp/js/main.js
@@ -1,82 +1,113 @@
 //refer http://www.smashingmagazine.com/2012/02/beginners-guide-jquery-based-json-api-clients/#the-full-code
-var IP = "localhost"
+var IP = "localhost";
 //var IP = "192.168.0.102"	//replace ip with localip If u wanna test from tab/mobile
-var rootURL = "http://" +IP +":8080/moviemon/movies";
-var loginURL = "http://" +IP +":8080/moviemon/authentication/signin"
-var favURL = "http://" +IP +":8080/moviemon/movies/favorites/"
-var watchListURL = "http://" +IP +":8080/moviemon/movies/watchlist/"
+var rootURL = "http://" + IP + ":8080/moviemon/movies";
+var loginURL = "http://" + IP + ":8080/moviemon/authentication/signin";
+var favURL = "http://" + IP + ":8080/moviemon/movies/favorites/";
+var watchListURL = "http://" + IP + ":8080/moviemon/movies/watchlist/";
 
 var allMovies;
 
-var unFavoriteIcon = "glyphicon glyphicon-star-empty gi-1x none"
-var favoriteIcon = "glyphicon glyphicon-star gi-1x gold"
+var unFavoriteIcon = "glyphicon glyphicon-star-empty gi-1x none";
+var favoriteIcon = "glyphicon glyphicon-star gi-1x gold";
 	
-var token = "authToken"
-
-
-var showPosters = function() {
-
-		var count = 0;
-	$.each(allMovies, function(index, el) {
-		
-		
-		var textToInsert = '';
-		
-		var genreString = '';
-		var genreLinks = '';
-		var genreList = el.genre.split(",");
-		for(i=0;i<genreList.length; i++)
-		{
-					genreString += '<a href="#">' + genreList[i] + '</a>' + ', '
-		}
-		genreLinks = genreString.substring(0,genreString.lastIndexOf(",")); //to remove the last extra comma. [ action, drama, ]
-		
-		var favIconToBeUsed;
-		if(el.favorite == true) {
-			favIconToBeUsed = favoriteIcon;
-		} else {
-			favIconToBeUsed = unFavoriteIcon;
-		}
-		count++;
-		
-		textToInsert += '<div class="col-lg-3  col-md-4 col-sm-6 col-xs-12"><div class="thumbnail fade"><img src="moviemon/posters/' + el.poster + '"  width="220" height="315">';
-		
-		textToInsert += '<a>' + '<span class="glyphicon glyphicon-bookmark gi-1x bookmark"' +'" onclick="tryWatchList(this , ' +el.id + ')"'   + 'title="Add to watchlist" data-wl-id=' + el.id +'></span></a>';
-		textToInsert += '<div class="caption">';
-		
-		
-		textToInsert += '<a>' + '<span class= "' + favIconToBeUsed + '" onclick="tryFavoriting(this , '+ el.id + ')" data-fav-id=' + el.id +'></a><span class="rating">' +' ' +el.imdbRating  +'<span class="ten">/10</span>' +'</span></span>';
-		
-		//textToInsert += '<a href="#">' + '<span class="glyphicon glyphicon-plus gi-2x"></span></a>';
-		
-		textToInsert += '<h3>' + el.title  + '<span id="year"> (<a href="">' + el.year +'</a>)</span>' +'</h3>';
-		textToInsert += '<div class="infoText">' + genreLinks +'<span class="pipe">' + ' | ' + '</span>' +el.runTime +'</div>';
-		
-		textToInsert +=  '</div></div></div>';
-		
+var wlistSelected = "glyphicon glyphicon-bookmark gi-1x bookmark";
+var wlistUnselected = "glyphicon glyphicon-bookmark gi-1x bookmark-none";
 	
-	//	if(el.favorite == true) {
-	//		$('#fav_section').append(textToInsert);
-	//	}
-	
-		$('#all_Movies').append(textToInsert);
-		
-        //if(count % 5 == 0) {
-		//	$('#watched_Movies').append(textToInsert);
-		//}
-		//testing purpose
-		if(count == 50) {
-			return false;
-		}
-
-	});
+var token = "authToken";
 
 
+var showPosters = function () {
+
+    var count = 0;
+    $.each(allMovies, function (index, el) {
+
+        var textToInsert = ''
+            , genreString = ''
+            , genreLinks = '';
+        var genreList = el.genre.split(",");
+        for (i = 0; i < genreList.length; i++) {
+            genreString += '<a href="#">' + genreList[i] + '</a>' + ', ';
+        }
+        genreLinks = genreString.substring(0, genreString.lastIndexOf(",")); //to remove the last extra comma. [ action, drama, ]
+
+        var favIconToBeUsed;
+        if (el.favorite === true) {
+            favIconToBeUsed = favoriteIcon;
+        } else {
+            favIconToBeUsed = unFavoriteIcon;
+        }
+        var wlistIcon;
+        if (el.inWatchList === true) {
+            wlistIcon = wlistSelected;
+        } else {
+            wlistIcon = wlistUnselected;
+        }
+
+        count++; //can be removed, just used to limit the number of movies. testing purpose
+
+        textToInsert += '<div class="col-lg-3  col-md-4 col-sm-6 col-xs-12"><div class="thumbnail fade"><img src="moviemon/posters/' + el.poster + '"  width="220" height="315">';
+
+        textToInsert += '<a>' + '<span class="' + wlistIcon + '" onclick="tryWatchList(this , ' + el.id + ')"' + 'title="Add to watchlist" data-wl-id=' + el.id + '></span></a>';
+
+        textToInsert += '<div class="caption">';
+
+        textToInsert += '<a>' + '<span class= "' + favIconToBeUsed + '" onclick="tryFavoriting(this , ' + el.id + ')" data-fav-id=' + el.id + '></a><span class="rating">' + ' ' + el.imdbRating + '<span class="ten">/10</span>' + '</span></span>';
+
+        //textToInsert += '<a href="#">' + '<span class="glyphicon glyphicon-plus gi-2x"></span></a>';
+
+        textToInsert += '<h3>' + el.title + '<span id="year"> (<a href="">' + el.year + '</a>)</span>' + '</h3>';
+        textToInsert += '<div class="infoText">' + genreLinks + '<span class="pipe">' + ' | ' + '</span>' + el.runTime + '</div>';
+
+        textToInsert += '</div></div></div>';
+
+        //	if(el.favorite == true) {
+        //		$('#fav_section').append(textToInsert);
+        //	}
+
+        $('#all_Movies').append(textToInsert);
+
+        //testing purpose
+        // if (count == 50) {
+        //   return false;
+        //    }
+
+    });
+
+
+};
+
+var setCookie = function (cname, cvalue) {
+    var d = new Date();
+    d.setTime(d.getTime() + (1 * 24 * 60 * 60 * 1000));
+    var expiry = "expires=" + d.toUTCString();
+    document.cookie = cname + "=" + cvalue + ";" + expiry;
+};
+
+function getCookie(cname) {
+    var name = cname + "=";
+    var ca = document.cookie.split(';');
+    for (i = 0; i < ca.length; i++) {
+        var c = ca[i];
+        while (c.charAt(0) == ' ') c = c.substring(1);
+        if (c.indexOf(name) == 0) return c.substring(name.length, c.length);
+    }
+    return "";
+};
+
+//set the http Authorization request header with token value
+//token value is retrieved from cookie
+function setHeader(xhr) {
+    var bearertoken = getCookie(token);
+    if (bearertoken) {
+        var bearerHeader = 'Bearer ' + bearertoken;
+        xhr.setRequestHeader('Authorization', bearerHeader);
+    }
 }
 
 function tryFavoriting(param, id) {
-	if (param.className == favoriteIcon) {
-		//Delete request to remove favorites
+	if (param.className === favoriteIcon) {
+    //Delete request to remove favorites
 		$.ajax({
 			url : favURL + id,
 			type : 'delete',
@@ -85,7 +116,7 @@ function tryFavoriting(param, id) {
 					param.className = unFavoriteIcon;
 				}
 			},
-			beforeSend: setHeader 
+			beforeSend: setHeader
 		});
 	} else {
 		//sending post request to add favorites
@@ -100,34 +131,34 @@ function tryFavoriting(param, id) {
 			beforeSend: setHeader 
 		});
 	}
-}
-//uncomment once watchlist field is sent in rest response !
+};
+
 function tryWatchList(param, id) {
-	/*if (param.className == favoriteIcon) {
-		//Delete request to remove favorites
+	if (param.className === wlistSelected) {
+		//Delete request to remove watchlist
 		$.ajax({
 			url : watchListURL + id,
 			type : 'delete',
 			statusCode : {
 				200 : function (response) {
-					param.className = unFavoriteIcon;
+					param.className = wlistUnselected;
 				}
 			},
 			beforeSend: setHeader 
 		});
-	} else { */
+	} else { 
 		//sending post request to add to watchlist
 		$.ajax({
 			url : watchListURL + id,
 			type : "POST",
 			statusCode : {
 				200 : function (response) {
-				//	param.className = favoriteIcon;
+					param.className = wlistSelected;
 				}
 			},
 			beforeSend: setHeader 
 		});
-	//}
+	}
 }
 
 var showScanButton = function () {
@@ -155,6 +186,8 @@ var getAllMovies = function () {
 				showPosters();
 			}
 			else {
+				document.getElementById("all_Movies").innerHTML = "";
+				//show no movies msg !
 				showScanButton();
 			}
 		},
@@ -175,6 +208,7 @@ var getFavMovies = function () {
 				showPosters();
 			}
 			else {
+				document.getElementById("all_Movies").innerHTML = "";
 				showScanButton();
 			}
 		},
@@ -196,6 +230,7 @@ var getWatchListMovies = function () {
 				showPosters();
 			}
 			else {
+				document.getElementById("all_Movies").innerHTML = "";
 				showScanButton();
 			}
 		},
@@ -222,34 +257,6 @@ function postLoginAjaxCall(username, password) {
 			    setCookie(token, data);
 		}
 	});
-}
-
-
-//set the http Authorization request header with token value
-//token value is retrieved from cookie
-function setHeader(xhr) {
-	var bearertoken = getCookie(token);
-	if(bearertoken) {
-		var bearerHeader = 'Bearer '+ bearertoken; 
-		xhr.setRequestHeader('Authorization', bearerHeader);
-	}
-}
-
-var setCookie = function(cname, cvalue) {
-    var d = new Date();
-    d.setTime(d.getTime() + (1*24*60*60*1000));
-    var expiry = "expires="+d.toUTCString();
-    document.cookie = cname +"=" + cvalue +";"+expiry;
-}
-function getCookie(cname) {
-    var name = cname + "=";
-    var ca = document.cookie.split(';');
-    for(var i=0; i<ca.length; i++) {
-        var c = ca[i];
-        while (c.charAt(0)==' ') c = c.substring(1);
-        if (c.indexOf(name) == 0) return c.substring(name.length,c.length);
-    }
-    return "";
 }
